Move all SqlTester Files into own folder
