+++ packages/database-editor/src/components/SqlQueryTester/SqlQueryContent.tsx
